requirements.txt
streamlit
PyPDF2
python-docx
docx2txt
pytesseract
pdf2image
pillow
